requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyotp==2.9.0
qrcode[pil]==7.4.2

# Caching & Background Tasks
redis==5.0.1
celery==5.3.4

# HTTP & Networking
httpx==0.25.2
aiofiles==23.2.1

# Data Processing
pandas==2.1.4
polars==0.20.3
numpy==1.25.2
scikit-learn==1.3.2
shap==0.43.0

# File Processing
python-magic==0.4.27
openpyxl==3.1.2

# Monitoring & Logging
structlog==23.2.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
ruff==0.1.6
bandit==1.7.5
safety==2.3.5